[Feature]: Moved our-service component to service folder
