+++ src/pages/service/index.jsx
@@ -1,5 +1,16 @@
+import OurServicePosterSection from './OurServicePosterSection';
+import { useEffect, useState } from 'react';
 function Service() {
-    return <div>Service</div>;
+    const [screenSize, setScreenSize] = useState(window.innerWidth);
+    useEffect(() => {
+        const handleResize = () => {
+            setScreenSize(window.innerWidth);
+        };
+        window.addEventListener('resize', handleResize);
+        return () => window.removeEventListener('resize', handleResize);
+    }, []);
+
+    return (<OurServicePosterSection screenSize={screenSize} />);
 }
 
 export default Service;
